Handle integers bigger than 64bit
The capacity of handling arbitrary size integers is quite important
to be able to replace the stdlib library.

Unfortunately RapidJSON isn't very flexible here.

For generating JSON it's quite easy to fallback to `Integer#to_s`.

However for parsing, the only escape hatch is to parse all
integers ourselves, which significantly impact performance.

It could be a bit better by parsing in place, but all Ruby functions
expect null terminated C-strings, so it wouldn't work well.

Ruby number parsing is also not as fast at RapidJSON's.

So for this to be performant, RapidJSON would need a flag that
allows to only parse over-sized integers ourself, and not all numbers.

But that would require a new feature upstream.