Raise an error if trying to serialize NaN or Infinity
It's debatable whether they should be supported. They are not valid
JSON according to RFC8259, however the stdlib JSON does parse them by
default and has an option not to.

RapidJSON also has a flag to parse them, but it's not set by this
gem.

So I think this gem should either parse and generate them, or
fail to do both, but right now it won't parse them but will
hapilly generated broken JSON when encoding them.