Refactor handling of non-native types
I believe that the flori/json behavior of calling either `to_json`
or `to_s` on unknown types is very much undesirable.

The first reason is that `Object#to_json` is defined, and fallback
as `Object#to_s`, meaning `JSON.dump()` accept absolutely anything
and will generally return valid JSON but garbage data.

I think it would be much better to explicitly define how types that
don't directly map to JSON are to be serialized.

The second reason is that `to_json` methods are global, so it's impossible
to serialize a type in one way in one part of the application, and in another
way elsewhere.

Typically you may way to serialize `Time` object in different formats when talking
to different APIs.

The solution in this PR is to provide a `Proc` to the encoder, when it
encounter an unknown type, it calls the provided proc and get either
a native type or a JSON fragment back.

This allows to both implement the `flori/json` API, but also to implement
some more restricted serialization using `case / when` or similar.