Revamp the JSONGem shim
This provide a JSON gem like interface for when you need
exact compatibility with it.

It silently fallback to calling the JSON gem if it receive
unknown options.

This can be used to selectively monkey patch code that would
benefit from a bit of extra speed.

Or it can be used directly, if the JSON gem semantic is desired.

The main `RapidJSON.load` / `RapidJSON.dump` now longer follow
the JSON gem semantic of calling `to_json` or `to_s` on unknown
objects.