Add json_ready? method
I'd hoped this could be done efficiently in pure Ruby, but the C version
is unfortunately still an order of magnitude faster.

An efficient implementation of json_ready? is necessary because it
allowed skipping a pass of deep-duping the hash structure passed in with
calls to as_json when the argument is already in a "json ready" format.